Modularize test harnesses for algorithms so that they can all be run from a single file. Will update data structures and create common test harness runner in future commit
