+++ DataStructures/LinkedList/test/LinkedListTest.cpp
@@ -8,17 +8,19 @@ int main(int argc, char** argv)
     return RUN_ALL_TESTS();
 }
 
+#include <string>
+
 TEST(LinkedListTest, append)
 {
-    LinkedList<int> ll;
+    LinkedList<string> ll;
 
-    ll.append(1);
-    ll.append(2);
-    ll.append(3);
+    ll.append("asdf");
+    ll.append("asdf");
+    ll.append("asdf");
 
-    EXPECT_EQ(ll.first()->data, 1);
-    EXPECT_EQ(ll.first()->next->data, 2);
-    EXPECT_EQ(ll.last()->data, 3);
+    // EXPECT_EQ(ll.first()->data, 1);
+    // EXPECT_EQ(ll.first()->next->data, 2);
+    // EXPECT_EQ(ll.last()->data, 3);
 }
 
 TEST(LinkedListTest, append_FollowingDeleteAll)
