+++ algorithms/LinkedList/ReverseTraversal/test/ReverseTraversalTestRunner.cpp
@@ -1,0 +1,7 @@
+#include "ReverseTraversalTestHarness.h"
+
+int main(int argc, char** argv)
+{
+    testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
