+++ algorithms/LinkedList/StraightTraversal/test/ForwardLinkedListTraversalTestRunner.cpp
@@ -1,0 +1,7 @@
+#include "ForwardLinkedListTraversalTestHarness.h"
+
+int main(int argc, char** argv)
+{
+    testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
