+++ algorithms/Search/BinarySearch/test/BinarySearchTestRunner.cpp
@@ -1,0 +1,9 @@
+#include "BinarySearchTestHarness.h"
+
+int main(int argc, char** argv)
+{
+    srand(time(nullptr));
+
+    testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
