+++ algorithms/Sets/PowerSet/test/PowerSetTestRunner.cpp
@@ -1,0 +1,7 @@
+#include "PowerSetTestHarness.h"
+
+int main(int argc, char** argv)
+{
+    testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
