+++ algorithms/Sorting/CountingSort/test/CountingSortTestRunner.cpp
@@ -1,0 +1,9 @@
+#include "CountingSortTestHarness.h"
+
+int main(int argc, char** argv)
+{
+    srand(time(nullptr));
+
+    testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
