+++ algorithms/Sorting/InsertionSort/test/InsertionSortTestHarness.h
@@ -1,22 +1,17 @@
-#include "../SelectionSort.h"
+#ifndef InsertionSortTestHarness_h
+#define InsertionSortTestHarness_h
+
 #include "../../../util/SortingUtil.h"
+#include "../InsertionSort.h"
 #include "gtest/gtest.h"
 
 #define ARRAY_SIZE 1000
 
-int main(int argc, char** argv)
-{
-    srand(time(nullptr));
-
-    testing::InitGoogleTest(&argc, argv);
-    return RUN_ALL_TESTS();
-}
-
-TEST(SelectionSort, ArraySort)
+TEST(InsertSort, ArraySort)
 {
     int arr[ARRAY_SIZE];
     size_t i = 0;
-    
+
     while (i < ARRAY_SIZE) {
         int randNum = rand() % ARRAY_SIZE;
 
@@ -26,12 +21,12 @@ TEST(SelectionSort, ArraySort)
         arr[i++] = randNum;
     }
 
-    selectionSort(arr, ARRAY_SIZE);
+    insertionSort(arr, ARRAY_SIZE);
 
     EXPECT_TRUE(arrayIsSorted(arr, ARRAY_SIZE));
 }
 
-TEST(SelectionSort, PointerArraySort)
+TEST(InsertSort, PointerArraySort)
 {
     int* arr[ARRAY_SIZE];
     size_t i = 0;
@@ -45,12 +40,12 @@ TEST(SelectionSort, PointerArraySort)
         arr[i++] = new int(randNum);
     }
 
-    selectionSort(arr, ARRAY_SIZE);
+    insertionSort(arr, ARRAY_SIZE);
 
     EXPECT_TRUE(pointerArrayIsSorted(arr, ARRAY_SIZE));
 }
 
-TEST(SelectionSort, VectorSort)
+TEST(InsertSort, VectorSort)
 {
     vector<int> arr(ARRAY_SIZE, 0);
 
@@ -63,7 +58,9 @@ TEST(SelectionSort, VectorSort)
         arr.push_back(randNum);
     }
 
-    selectionSort(arr);
+    insertionSort(arr);
 
     EXPECT_TRUE(vectorIsSorted(arr));
 }
+
+#endif
