+++ algorithms/Sorting/SelectionSort/test/SelectionSortTestRunner.cpp
@@ -1,0 +1,9 @@
+#include "SelectionSortTestHarness.h"
+
+int main(int argc, char** argv)
+{
+    srand(time(nullptr));
+
+    testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
