+++ algorithms/Strings/HammingDistance/test/HammingDistanceTestRunner.cpp
@@ -1,0 +1,7 @@
+#include "HammingDistanceTestHarness.h"
+
+int main(int argc, char** argv)
+{
+    testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
+}
